app/requirements.txt
boto3==1.34.0
psycopg2-binary==2.9.9